pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mac-calendar-mcp"
version = "0.1.0"
description = "MCP server providing access to macOS Calendar events (including synced Google Calendar)"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["mcp", "calendar", "macos", "eventkit", "google-calendar"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "mcp>=0.9.0",
    "pyobjc-framework-EventKit>=10.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "freezegun>=1.2.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/mac-calendar-mcp"
Repository = "https://github.com/yourusername/mac-calendar-mcp"
Issues = "https://github.com/yourusername/mac-calendar-mcp/issues"

[project.scripts]
mac-calendar-mcp = "mac_calendar_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mac_calendar_mcp"]
